# docs/technical/api-documentation.yaml
openapi: 3.0.3
info:
  title: Post-Facto Coding Review API
  description: |
    HIPAA-compliant API for automated medical coding review and revenue optimization.

    ## Overview
    This API allows healthcare providers to submit clinical notes and billed codes for automated review,
    identifying missed billing opportunities using AI-powered analysis.

    ## Authentication
    All API endpoints require JWT authentication via Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## HIPAA Compliance
    All PHI is processed through Amazon Comprehend Medical for de-identification before AI analysis.
    Data is encrypted at rest and in transit (TLS 1.3).

  version: 1.0.0
  contact:
    name: RevRX Support
    email: support@revrx.com
  license:
    name: Proprietary
servers:
  - url: https://api.revrx.com/v1
    description: Production server
  - url: https://staging-api.revrx.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server
...
paths:
...
  /encounters/upload-note:
    post:
      tags:
        - Encounters
      summary: Upload clinical note
      description: |
        Upload a clinical note file (TXT, PDF, or DOCX format).
        Maximum file size: 5MB.
        Creates a new encounter record.
      operationId: uploadNote
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Clinical note file
                patientAge:
                  type: integer
                  minimum: 0
                  maximum: 150
                patientSex:
                  type: string
                  enum: [M, F, O]
                visitDate:
                  type: string
                  format: date
                encounterType:
                  type: string
                  example: Office Visit
      responses:
        '201':
          description: Note uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncounterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          description: File too large
        '415':
          description: Unsupported file type
  /encounters/{encounterId}/upload-codes:
    post:
      tags:
        - Encounters
      summary: Upload billing codes
      description: Upload billed CPT/ICD codes in CSV or JSON format
      operationId: uploadBillingCodes
      parameters:
        - name: encounterId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Billing codes file (CSV or JSON)
          application/json:
            schema:
              type: object
              required:
                - codes
              properties:
                codes:
                  type: array
                  items:
                    $ref: '#/components/schemas/BillingCode'
      responses:
        '200':
          description: Billing codes uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncounterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /encounters:
    get:
      tags:
        - Encounters
      summary: List encounters
      description: Get list of user's encounters with pagination
      operationId: listEncounters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, complete, failed]
      responses:
        '200':
          description: Encounters retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  encounters:
                    type: array
                    items:
                      $ref: '#/components/schemas/EncounterResponse'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EncounterResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        uploadDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, processing, complete, failed]
        processingTime:
          type: number
          description: Processing time in seconds
        patientAge:
          type: integer
        patientSex:
          type: string
        visitDate:
          type: string
          format: date
        encounterType:
          type: string
    BillingCode:
      type: object
      required:
        - code
        - type
      properties:
        code:
          type: string
          example: "99214"
        type:
          type: string
          enum: [CPT, ICD10]
        description:
          type: string
...
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        itemsPerPage:
          type: integer
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []